# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Nemo V2
  version: 1.0.0
...
paths:
  /api/v1/platform/modules:
    get:
      tags:
        - FMS JS > Platform
      summary: Get Console Modules
      description: Get all platform modules available for console access
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Console permissions fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      userid:
                        type: string
                        description: User ID
                      modules:
                        type: array
                        description: Array of platform modules
                        items:
                          type: object
                          properties:
                            moduleid:
                              type: string
                              description: Module ID
                            moduleinfo:
                              type: object
                              description: Module information including name, type, and other details
                              properties:
                                moduleid:
                                  type: string
                                  description: Module ID
                                modulename:
                                  type: string
                                  description: Module name
                                moduletype:
                                  type: string
                                  description: Module type (e.g., platform)
                                modulecode:
                                  type: string
                                  description: Module code
                                moduleinfo:
                                  type: object
                                  description: Additional module information
                                creditspervehicleday:
                                  type: number
                                  description: Credits per vehicle per day
                                isenabled:
                                  type: boolean
                                  description: Whether module is enabled
                  msg:
                    type: string
                    example: "Console permissions fetched successfully"
        "400":
          description: Bad Request - Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "INPUT_ERROR"
                      errdata:
                        type: string
                        example: "Invalid User ID format"
                  msg:
                    type: string
                    example: "Invalid input data"
        "403":
          description: Forbidden - Insufficient permissions
        "500":
          description: Internal Server Error - Get console permissions failed
...
  /api/v1/platform/module:
    post:
      tags:
        - FMS JS > Platform > Module
      summary: Create Module
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                modulename: MyTripInsights
                moduletype: web
                modulecode: mytripinsights
                creditspervehicleday: 5
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/platform/role:
    post:
      tags:
        - FMS JS > Platform > Roles
      summary: Create Role
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                rolename: Customer Care Admin L3
                roletype: platform
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/platform/pkg:
    post:
      tags:
        - FMS JS > Platform > Packages
      summary: Create Default Package
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                pkgname: Pro
                pkgtype: default
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/platform/home:
    get:
      tags:
        - FMS JS > Platform
      summary: Get Console Home
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: INVALID_TOKEN
                      errdata:
                        type: string
                        example: Invalid token
                  data:
                    type: "null"
...